+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:npng/generated/l10n.dart';
 import 'package:npng/route_map.dart';
-import 'package:npng/theme_material.dart';
+import 'package:npng/theme.dart';
 import 'package:provider/provider.dart';
 
 import 'package:npng/data/models/workout_provider.dart';
@@ -13,14 +13,13 @@ import 'package:npng/data/repository.dart';
 
 import 'package:npng/data/sqlite/sqlite_repository.dart';
 
-//final bool isApple = !kIsWeb && (Platform.isMacOS || Platform.isIOS);
-const bool isApple = false;
-
+// final bool isApple = !kIsWeb && (Platform.isMacOS || Platform.isIOS);
 // bool get isMobileDevice => !kIsWeb && (Platform.isIOS || Platform.isAndroid);
-final bool isDesktopDevice =
-    !kIsWeb && (Platform.isMacOS || Platform.isWindows || Platform.isLinux);
 // bool get isMobileDeviceOrWeb => kIsWeb || isMobileDevice;
 // bool get isDesktopDeviceOrWeb => kIsWeb || isDesktopDevice;
+
+final bool isDesktopDevice =
+    !kIsWeb && (Platform.isMacOS || Platform.isWindows || Platform.isLinux);
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -45,7 +44,7 @@ class Application extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      debugShowCheckedModeBanner: false,
+      //debugShowCheckedModeBanner: false,
       localizationsDelegates: const [
         S.delegate,
         GlobalMaterialLocalizations.delegate,
@@ -54,9 +53,9 @@ class Application extends StatelessWidget {
       ],
       supportedLocales: S.delegate.supportedLocales,
       onGenerateTitle: (BuildContext context) => S.of(context).title,
-      //theme: AppTheme.light,
       theme: AppTheme.light,
       darkTheme: AppTheme.dark,
+      themeMode: ThemeMode.system,
       initialRoute: kInitialRoute,
       routes: appRoutes,
     );