+++ lib/screens/log/log_calendar_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:npng/data/models/models.dart';
 import 'package:npng/data/repository.dart';
 import 'package:npng/screens/log/log_show_workout_screen.dart';
-import 'package:npng/theme_material.dart';
+import 'package:npng/theme.dart';
 import 'package:provider/provider.dart';
 
 class LogCalendarScreen extends StatefulWidget {