+++ lib/theme.dart
@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
 
+const Color kActionColorIcon = Colors.white;
+const Color kActionColorEdit = Colors.grey;
+const Color kActionColorDelete = Colors.black;
+
 class AppTheme {
   /// Material light theme
   static ThemeData light = ThemeData(