+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 2.1.0
 - Same desingn in all operation systems.
 - Excersises checkbox on workout screen.
-- Program selection now is on the Workout screen appbar.
+- Program selector now is on the Workout screen appbar.
 - Removed window header in macOS.
+- Improved work with small screens.
 - Improved perfomance.
 
 ## 2.0.1